Remove quotes around STL module names
This commit introduces an extra state to the STL scanner in order to
detect if the current identifier is a module name. If it is, the
surrounding quotes are cut. The quotes are syntactically required but do
not contribute to the module names being unique, therefore cutting them
is allowed. Doing so simplifies both the user and the language
interface. Existing regression tests were changed in a way that all
directly refer to the default entry point which was introduced by a
past commit. One test, Function_Call2, is now responsible for checking
the --function flag.