Fix primitive wrapper types in JSON value retriever
Following the format of the json-io library, the JSON input to the
value retriever does not distinguish between primitive types and their
corresponding wrapper types. A class Foo with a field fooField of type
Integer might appear as
{
  "@type":"org.cprover.Foo",
  "fooField":100
}
To support this "primitive representation", we need a special case in
the struct assignment part.