Refactor and tidy how function calls are handled in AI
The current ai_baset has a "skip if not available, otherwise recurse"
policy which cannot be altered or removed.  Recursion causes problems
for domains with large height as they can crash the analyser.  It also
mixes the (interprocedural) state of the /analysis/ with the recursive
state of the /analyser/.  This forces an undesireable design choice --
it means that anything interprocedural must be handled purely in the
domain.  Doing so has several limitations and so this patch allows
inheritance to change how interprocedural analysis is done.