goto-symex: Reformat parameter conversion error message
Language front-ends cannot always type check parameter conversions, for
example in case of incomplete forward declaration there isn't sufficient
information available. Thus some of that work is left to be done during
symbolic execution. In such cases, late type checking can fail and the
problem needs to be reported back to the users. We previously did not
include the source location of the problem in the output, and did not
make generous use of newlines, making diagnosing the problem difficult.