Check for unreachability before true / false in AI verification
There will be a few cases that were previously marked as either
"SUCCESS" or "FAILURE (if reachable)" that will now be marked as
"SUCCESS (unreachable)".  In the first case this is definitely an
improvement as it is more precise and generally people want to
know when assertions are safe just because they are unreachable.
In the second case it is a bit more debateable but I think that
saying "FAILURE (if reachable)" is kind of disengenuous when we
know that it isn't reachable.