Fix the ordering of variable initialisation
by linearising the preorder induced by the symbol values referring to other
symbols.

Previously we were using the `refers-to` relation (`a` refers to `b` if the
value `a` should be initialised with contains -- as a subexpression -- the
symbol `b`) as if it was a linear ordering between the snapshot symbols. That
is, we called std::stable_sort with this relation which is incorrect.

Now we select one linearisation of the preorder (given by `refers-to`) to be the
initilisation ordering. Note that we ignore cycles, or more precisely in the
presence of a cycle we start from the first visited symbol.