Test for complexity limits
This test will fail if any of the features this relies upon has been
broken, but has the side-affect that if things ARE broken it'll take
a long time to fail. Not much to do about that, as this sort of
long-running symex-destroying code is what this feature is meant for.