Revamp function contracts interface
This commit adds four new flags to goto-instrument that allow CBMC to
check or take advantage of function contracts in the target codebase.
It prepares for a series of commits that will improve the soundness of
this functionality and add support for loop invariants.

The motivation behind this patch series is to improve scalability of
code verification. Given the following code:

  foo()
  __CPROVER_ensures(P)
  __CPROVER_ensures(Q)
  { ... }

  bar()
  {
    foo();
  }

the user can pass `--replace-call-with-contract foo` to goto-instrument,
which will change the body of `bar` so that it can be verified more
scalably:

  bar()
  {
    assert(P);
    assume(Q);
  }

The user should then separately pass `--enforce-contract foo` to
goto-instrument. This instruments `foo` so that it looks like this:

  foo()
  {
    assume(P);
    // original body
    assert(Q);
  }

Running `cbmc --function foo` on the instrumented binary then checks
that `foo` actually abides by its contract, regardless of its calling
context.

Users can also pass `--enforce-all-contracts` and
`--replace-all-calls-with-contracts` to avoid naming every function.