Fix string solver string-to-int conversion
Previously this both codified the relationship between strings and integers, *and* constrained the
input string to only contain appropriate characters. The constraint part of that is incorrect: it is
not functional (it has side-effects on its input), which meant the input string was also constrained
in other contexts. "if(s.equals("Hello")) { return 1; } else { return Integer.parseInt(s); }", for
example, would be partly unsatisfiable because s is constrained to only consist of digits due to the
parseInt operation.

This is fixed by splitting is_valid_int from parse_int, and updating models libraries to make the
validation check and throw a NumberFormatException if necessary, before using the now-pure-functional
parse_int routine