Track hardness of solver queries
Introduces a functionality to track and collect the complexity of individual
solver queries. For SAT solvers we collect the number of clauses, literals, and
variables. The functionality matches solver queries to the SSA steps that
produced them. After running the solver, we produce a JSON report that matches
program line to GOTO instruction to SSA step to solver hardness.

At present the collection is command-line optional and only works with the
default SAT solver (minisat2). Extension to other SAT and SMT solvers should be
possible.

A simple regression test is included.

Note: the solver-hardness structure lives in the minisat2 and is accessible via
the interface of hardness-collector (which minisat2 inherits from). Hence to
find out if symex should collect the statistics one must try-cast `prop` to
`hardness-collector`. `prop` is part of the `solvert` but not directly
accessible from `decision-procedure`. So we also had to extend the interface
that initialises the SSA conversion (to include `prop` as well).