Only enable location context if value set is turned off
There is currently some inconsistency between how the value set and the
interval domains handle wrapping. Specifically, the value set returns an
unwrapped object when eval'd, whereas the interval domain returns a
wrapped object. Since the rest of the code is compatible with the
interval suggestion, we temporarily disable wrapping if using value set