Add support for invoke-virtual and invoke-interface lambdas
To support these we store lambda descriptors in the same form as virtual method references,
using class_method_descriptor_exprt, lowering it to a symbol_exprt if we have a known callee.

This is also the first form of synthetic method to produce a virtual method call, hence the
alteration of notify_static_method_calls to tolerate them.