Add a caller_history argument to ai_historyt::step()
This is similar to a three-way merge in that it allows histories to
not only use their state and histories in the next location but also
to access the history at the call site.  Doing so makes it possible to
create "function local" history where all call-sites merge and drop
their previous history but it is picked back up on function return.
This is needed for managing scalability in some of the more involved
histories.