Avoid unintentional copy in language_file.h
The iterators for `lazy_method_map` return
`std::pair<const irep_idt, language_filet *>` (notice the const). Because of
this type mismatch in the iterator there is a forced copy here.

Recent versions of clang warn about this and thus prevent building with -Wall
-Werror.

Worth noting we don’t actually use the remove and clear methods in this class so
we could probably also just remove them.