Remove multiple function assertion
This commit removes an assertion that triggers when multiple functions
get mangled to the same name when using the --export-file-local-symbols
flag to goto-cc.

This assertion was previously valid when compiling files one at a time, but
commit 429313e introduced the possibility of compiling and linking several
files at once. This is a problem when two source files both include an
implementation file, like [1], because the implementation may contain static
functions that get mangled twice.

This commit replaces the assertion with a debug message in case users would
like to check how often this happens.

This commit fixes #5380.

[1] https://github.com/awslabs/aws-c-common/blob/master/include/aws/common/atomics_gnu.inl