github: always test Xen
We want to make sure Xen related workflows are not broken in the future.
As github workflows offer 6h runtime, this is a good opportunity.
Furthermore, setup and usage is simple.

The current state of the file adds compiling CBMC, as well as getting
Xen. We will add next steps incrementally.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>

CR: https://code.amazon.com/reviews/CR-28954277