Remove dead method from recursive initialization
The ::build_array_string_constructor method was originally intended to be used
for building C string constructors. Unfortunately as written the condition for
calling it was actually impossible to satisfy. This wasn’t noticed so far
because similar code is now generated in ::build_dynamic_array_constructor
instead.

Dead code is a maintenance burden so it’s removed here.