Generate index expressions for deref(p + offset), where p points to an array in a struct
Currently if offset is unknown, value_set_dereference will resolve p + offset to [(o1, unknown), (o2, unknown), ...], ultimately leading to a byte-update operation that may touch any of o1, o2 etc and specifically knocks them both out of the constant propagator.

With this change if p is known to point to arrays or array-members of structs then we will produce an expression like (c1 ? array1[offset] : c2 ? someobj.array2[offset] : ...)

Field-sensitivity may then kick in and update only the array within someobj, rather than the whole struct (this is unsound, but mirrors our behaviour when an array-in-struct is directly referenced. We may want to make this behaviour optional if the user is intentionally searching for solutions involving out-of-bounds accesses)

Note this is quite similar to value_set_dereferencet::build_reference_to's behaviour when referring to an array (but not an array within a struct), which builds an index expression instead of a byte-extract op. We can't extend that code however as the (constant + nonconstant) form of the pointer being dereferenced is lost by then.

Fixes: #5397