Add more robust type name to type identifier conversion
In order to certain that we don't emmit C harnesses containing invalid
identifiers.

The unit test includes both a function to generate the string of all
printable characters and the same string as a string literal in order to
show that the string does include all printable characters and to show
what these characters are.