Add binary representation to all bitvector values in the xml trace
The purpose of this is to provide a consistent interface for getting the
values of bitvector-ed types out of the trace.

An existing test is updated in this commit in order for the existing
to test to match new functionality and so that every commit still passes
all tests.

There are instances when running with `--cprover-smt2` where the type of
the expression is `bitvector_typet`, but the expression is not a
`constant_exprt`. By removing the printing in this case we can avoid
breaking the partial functionality available with the `--cprover-smt2`
option.