Delegate copy command to cmake binary.
The CMakeLists.txt was performing a `cp` command
directly, which while it was not an issue on linux,
if you build on Windows with cmake, it fails the
build because the equivalent command is `copy`.

I made it so we use the portable interface of
the CMake binary to perform the copy across all
targets, so it's portable.