Refactor `prefix_in_list` to take vector of strings
Because this gives us compatability with modern C++ features such as
ranged for loops and the algorithms in the standard template library.
Because using a standard container rather than a null terminated array
improves memory safety.

Clang format is off for the lists of options in order to keep the
existing manual formatting. This existing formatting is consistent with
the other lists of options.